Fix failed schlieren directory bug
Collapsed reading in of schlieren and sensor directories into a single
step, rather than storing them in a local variable as well as the
dataframe. Storing them separately caused a bug when the local variable
was referenced after the failed shots had been filtered out of the
dataframe, resulting in the wrong directory being stored.