Fix calibration functions
* correctly apply pixel uncertainty
* correctly apply line thickness uncertainty
* add functionality to exit calibration imshow by hitting 'enter' key

Note: for unknown reasons, jupyter started having freezing problems with the
calibration process. The problem happens on line 173, which is where the
process seems to hang up. This is a different issue than the previous
freezing issue, which was fixed by adding block=True, and manifests
itself as an interminable execution hold after exiting the figure rather
than freezing on figure instantiation. IDK what's up here, but I'm going
to just stop using jupyter notebooks for the spatial calibrations and
shift everything into a .py file.