Fix sdtoolbox crash problem
Wrapped cvsolve calls in a try/except block since I found that sdtoolbox
was crashing due to me not sending it a long enough simulation time. The
time required is mixture-dependent, so now it should try 10 times,
doubling in length each iteration, before it gives up the ghost and
crashes. There may be a better way to predict the final sim time, and I
will leave that until this stops working.