Fix bugs caused by lack of diode tdms
Lack of diode tdms leads to NaN values being passed in place of a string
for diode data location, which caused a mess of issues. These have been
cleaned up, and velocity calculation now returns NaN+/-NaN on a failed
calculation, which it probably should have done in the first place
instead of 0+/-0.